修改了设置变量值的逻辑，现在可以修改为double类型

+++ ExprStruct.cpp
@@ -85,7 +85,7 @@ QString Expr::toInfix() const
 }
 
 
-void Expr::setVarValue(char c,int v){
+void Expr::setVarValue(char c,double v){
     varMap[c] = v;
 }
